updated login/signup body validation error messages to match auth me and not api doc

+++ backend/db/seeders/20231002004855-demo-memberships.js.js
@@ -40,12 +40,10 @@ module.exports = {
 
   async down (queryInterface, Sequelize) {
 
-    await queryInterface.bulkDelete(options)
-
-    // await queryInterface.bulkDelete(options, {userId: membershipsData.map(data => data.userId),
-    //   groupId: membershipsData.map(data => data.groupId),
-    //   status: membershipsData.map(data => data.status)
-    // })
+    const Op = Sequelize.Op;
+await queryInterface.bulkDelete(options, {userId: membershipsData.map(data => data.userId),
+      groupId: {[Op.in]: membershipsData.map(data => data.groupId)},
+    }, {});
 
   }
 };
